% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{lFetchDailyHistoricalFxRates}
\alias{lFetchDailyHistoricalFxRates}
\title{Fetch NBP rates against indicated foreign currencies}
\usage{
lFetchDailyHistoricalFxRates(cTickers, dateStartDate, dateEndDate)
}
\arguments{
\item{cTickers}{- character vector of IDs of the}

\item{dateStartDate}{- Date class scalar, indicates the date of the beginning
of the interval for which the data is to be fetched. It is included in the
final interval.}

\item{dateEndDate}{- Date class scalar, indicates the date of the end
of the interval for which the data is to be fetched. It is included in the final interval.}
}
\value{
list of data.tables; names of the data.tables are standard ISO FX
pairs names with PLN featuring as the quote currency and the currencies indicated
in cTickers as base currencies; e.g. the results for query of USD rates
will be available under the output list's key "USDPLN".
}
\description{
Fetch FX rates against given foreign currencies from NBP API
}
\details{
This function fetches the FX rates against Polish zloty for the currencies
indicated in the input character vector cTickers. The data is fetched for the
time scope defined by the parameters dateStartDate and dateEndDate
}
